TEST: Recover two test files in s2plib and add them to unittest
